+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* Ability to filter statedb job queries using a specific
+  beginswith string
+
 
 ## [v0.7.0] - 2022-02-17